+++ CHANGELOG.md
@@ -377,3 +377,11 @@
 ### version 7.8 MARLO BARUA
 
 - Partially Done with AdminHR and AdminOSP(Main Dashboard, Archive, View Details, and Edit )
+
+### version 7.9 MARLO BARUA
+
+- Partially Done with AdminCBSS(Main Dashboard, Archive, View Details, and Edit )
+
+### version 7.10 KIARA HAGUNOB
+
+- Partially fixed CBSS, and added border radius in SWDA